Refactor the core random reader to be in the `tpm` package
